asiocurl::io_service::do_action - Improve Exception Support
asiocurl::io_service::do_action (a private method) is no longer noexcept.  As a result of this the callbacks passed to Boost.ASIO are no longer noexcept.  Exceptions thrown from asiocurl::io_service::do_action propagate up the stack and are rethrown from boost::asio::io_service::poll, ::poll_one, ::run, or ::run_one.  This does not apply to exceptions thrown by asiocurl::io_service::do_action when it is called from asiocurl::io_service::timer when it is invoked with timeout_ms less than or equal to zero: In this case the invocation of asiocurl::io_service::do_action is not asynchronous and therefore the exception is simply swallowed and an error is returned to libcurl.