asiocurl::io_service - libcurl Possible Bug Workaround
It appears that libcurl may invoke the multi handle's "socket function" with CURL_POLL_REMOVE after it calls the appropriate easy handle's "close function".  This may be a bug in libcurl as this sequence of events may introduce a race condition (it makes the assumption that the socket file descriptor is not rapidly reused by the operating system/application).

This was leading to a segmentation fault/undefined behaviour since asiocurl::io_service::socket (i.e. the "socket function") assumed that the provided socket was valid (i.e. had not been closed).