src/configure.hpp.in
Added the configure.hpp.in file which is preprocessed by CMake into the include/asiocurl/configure.hpp file based on the build options.  This allows for the value of USE_BOOST_FUTURE to be preserved when included by client code rather than putting the onus of defining (or not) ASIOCURL_USE_BOOST_FUTURE on them.